+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Support ignoring, making, copying and renaming directories and symlinks
 - Add option to expand variable values
 - Add parenthesis to expressions
+- Add equals to expressions
 
 ## v1.1.1
 - Prepare for AUR release